Split config in separated files for different environments
