+++ config/webpack.config.react.js
@@ -1,0 +1,41 @@
+/* eslint-disable import/no-extraneous-dependencies */
+const { resolve } = require('path');
+const webpack = require('webpack');
+const StyleLintPlugin = require('stylelint-webpack-plugin');
+const ExtractTextPlugin = require('extract-text-webpack-plugin');
+/* eslint-enable import/no-extraneous-dependencies */
+
+const entries = {
+  app: `${resolve(__dirname, '../src')}/main.js`,
+  vendor: ['react', 'redux', 'react-dom'],
+};
+
+module.exports = {
+  entry: entries,
+  devtool: 'source-map',
+  devServer: {
+    contentBase: 'src',
+    inline: true,
+    port: 8080,
+    historyApiFallback: true,
+  },
+  plugins: [
+    new StyleLintPlugin({
+      context: `${resolve(__dirname, '../src')}/`,
+      files: '**/*.css',
+      config: {
+        extends: 'stylelint-config-standard',
+        rules: {
+          'selector-pseudo-class-no-unknown': null,
+          'unit-whitelist': ['px', 'deg', '%', 'em', 'ms'],
+          'length-zero-no-unit': null,
+        },
+        ignoreFiles: './node_modules/**/*.css',
+      },
+    }),
+    new ExtractTextPlugin({
+      filename: 'styles.css',
+      allChunks: true,
+    }),
+  ],
+};
